pyproject.toml
[project]
name = "python-project-template"
version = "0.1.0"
description = ""
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-openai>=0.3.4",
    "nemoguardrails>=0.11.1",
    "openai>=1.61.1",
    "python-dotenv>=1.0.1",
]

[tool.mypy]
strict = true

[tool.ruff.lint]
select = ["ALL"]
ignore = ["D", "ANN", "TRY", "EM", "FA", "SIM108", "S101", "RET505", "TCH002"]
